Added namespaces to the Messaging Code, and removed hardcoding
